requirements.txt
Flask
Flask-Cors
Gunicorn
openai
python-dotenv

firebase-admin